# Conflicts:
#	src/main/java/net/playtogether/jpa/controller/ChampionshipController.java
#	src/main/java/net/playtogether/jpa/repository/TeamRepository.java

Arreglados fallos de merge con invitación a torneos, implementada
seguridad en sistema de invitación a torneos y arreglados tests de
invitaciones

+++ src/main/java/net/playtogether/jpa/controller/InvitationController.java
@@ -1,6 +1,7 @@
 
 package net.playtogether.jpa.controller;
 
+import java.security.Principal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -21,9 +22,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import net.playtogether.jpa.entity.Invitation;
 import net.playtogether.jpa.entity.Team;
-import net.playtogether.jpa.entity.User;
+import net.playtogether.jpa.entity.Usuario;
 import net.playtogether.jpa.service.ChampionshipService;
 import net.playtogether.jpa.service.InvitationService;
+import net.playtogether.jpa.service.UsuarioService;
 
 @Controller
 public class InvitationController {
@@ -34,6 +36,9 @@ public class InvitationController {
 	@Autowired
 	ChampionshipService championshipService;
 
+	@Autowired
+	UsuarioService userService;
+
 
 	@InitBinder("invitation")
 	public void initInvitationBinder(final WebDataBinder dataBinder) {
@@ -42,8 +47,9 @@ public class InvitationController {
 	
 
 	@GetMapping("/invitations/team/{teamId}")
-	public String searchParticipants(final ModelMap model, @PathVariable("teamId") final Integer teamId) {
+	public String searchParticipants(final ModelMap model, @PathVariable("teamId") final Integer teamId, Principal principal) {
 		Team team = this.championshipService.findTeamId(teamId);
+		
 		model.put("team_participants", team.getParticipants());
 		model.put("teamSize", team.getTeamSize());
 		model.put("team", team);
@@ -54,12 +60,21 @@ public class InvitationController {
 
 	@GetMapping("/invitations/team/{teamId}/send_invitation")
 	public String initSendInvitation(@RequestParam(value = "search", required = false) final String search,
-			final ModelMap model, @PathVariable("teamId") final Integer teamId) {
-		List<User> searched_users = new ArrayList<>();
+			final ModelMap model, @PathVariable("teamId") final Integer teamId, Principal principal) {
+		List<Usuario> searched_users = new ArrayList<>();
 		Team team = this.championshipService.findTeamId(teamId);
 		model.put("team_participants", team.getParticipants());
 		model.put("teamSize", team.getTeamSize());
 		model.put("team", team);
+		
+		if (!team.getUser().getUser().getUsername().equals(principal.getName())) {
+			model.put("loggedUserIsNotTheTeamOwner", true);
+			return "invitations/addParticipantsForm";
+		}
+		
+		if (search == null || search.equals("")) {
+			return "invitations/addParticipantsForm";
+		}
 
 		if (team.getParticipants().size() >= team.getTeamSize()) {
 			model.put("limitedTeamSize", true);
@@ -72,34 +87,47 @@ public class InvitationController {
 			model.put("notMoreUsers", true);
 			return "invitations/addParticipantsForm";
 		}
+		
+		Usuario user = this.userService.findByUsername(principal.getName());
+		if (searched_users.contains(user)) {
+			searched_users.remove(user);
+		}
+		
 		model.put("searched_users", searched_users);
 		return "invitations/listSearchedUsers";
 	}
 
 	@PostMapping("/invitations/team/{teamId}/send_invitation")
 	public String postSendInvitation(@ModelAttribute("selected_participant") final String selected_participant,
-			@PathVariable("teamId") final int teamId, final BindingResult result, final ModelMap model) {
+			@PathVariable("teamId") final int teamId, final BindingResult result, final ModelMap model, Principal principal) {
 		
-		User participant = this.championshipService.findUsersById(Integer.parseInt(selected_participant));
+		Usuario participant = this.championshipService.findUsersById(Integer.parseInt(selected_participant));
 		Team team = this.championshipService.findTeamId(teamId);
-		
-		if (!team.getChampionship().getFinishDate().isBefore(LocalDate.now())) {
-			if (isNotInChampionshipTeam(team.getChampionship().getId(), participant)) {	
-				Boolean isNotInvitedYetToChampionshipTeam = this.invitationService.isNotInvitedYetToChampionshipTeam(teamId, participant.getId());			
-				
-				if (isNotInvitedYetToChampionshipTeam) {
-					Invitation invitation = new Invitation();
-					invitation.setTeam(team);
-					invitation.setReceiver(participant);
-					this.invitationService.save(invitation);
+		if (team.getUser().getUser().getUsername().equals(principal.getName())) {
+			if (!principal.getName().equals(participant.getUser().getUsername())) {
+				if (!team.getChampionship().getFinishDate().isBefore(LocalDate.now())) {
+					if (isNotInChampionshipTeam(team.getChampionship().getId(), participant)) {	
+						Boolean isNotInvitedYetToChampionshipTeam = this.invitationService.isNotInvitedYetToChampionshipTeam(teamId, participant.getId());			
+						
+						if (isNotInvitedYetToChampionshipTeam) {
+							Invitation invitation = new Invitation();
+							invitation.setTeam(team);
+							invitation.setReceiver(participant);
+							this.invitationService.save(invitation);
+						} else {
+							model.put("alreadyInvited", true);
+						}
+					} else {
+						model.put("alreadyInChampionshipTeam", true);
+					}
 				} else {
-					model.put("alreadyInvited", true);
+					model.put("championshipIsFinished", true);
 				}
 			} else {
-				model.put("alreadyInChampionshipTeam", true);
+				model.put("cantSelfInvite", true);
 			}
 		} else {
-			model.put("championshipIsFinished", true);
+			model.put("loggedUserIsNotTheTeamOwner", true);
 		}
 		model.put("team_participants", team.getParticipants());
 		model.put("teamSize", team.getTeamSize());
@@ -107,16 +135,16 @@ public class InvitationController {
 		return "invitations/addParticipantsForm";
 	}
 
-	private Boolean isNotInChampionshipTeam(Integer championshipId, User user) {
+	private Boolean isNotInChampionshipTeam(Integer championshipId, Usuario user) {
 		Boolean isNotInChampionshipTeam = this.championshipService.findTeamsByChampionshipId(championshipId)
 				.stream().flatMap(t -> t.getParticipants().stream()).noneMatch(u -> u.equals(user));
 		return isNotInChampionshipTeam;
 	}
 	
-	private void deleteRepeatedUsers(final Team team, final List<User> searchedUsers) {
-		for (int i = 0; i < searchedUsers.size(); i++) {
-			if (team.getParticipants().contains(searchedUsers.get(i))) {
-				searchedUsers.remove(i);
+	private void deleteRepeatedUsers(final Team team, final List<Usuario> searched_users) {
+		for (int i = 0; i < searched_users.size(); i++) {
+			if (team.getParticipants().contains(searched_users.get(i))) {
+				searched_users.remove(i);
 				i--;
 			}
 		}
@@ -124,9 +152,8 @@ public class InvitationController {
 	
 	
 	@GetMapping("/invitations/championshipInvitations")
-	public String listChampionshipInvitations(final ModelMap model) {
-		Collection<Invitation> invitations = this.invitationService.findChampionshipInvitationsByUserName("Antonio");	// CAMBIAR AL NAME DEL PRINCIPAL CUANDO SE PUEDA
-		
+	public String listChampionshipInvitations(final ModelMap model, Principal principal) {	
+		Collection<Invitation> invitations = this.invitationService.findChampionshipInvitationsByUsername(principal.getName());		
 		List<Invitation> expiredInvitations = invitations.stream().filter(i -> i.getTeam().getChampionship().getFinishDate().isBefore(LocalDate.now())).collect(Collectors.toList());
 		
 		if (expiredInvitations.size() > 0) {
@@ -141,12 +168,26 @@ public class InvitationController {
 
 	@GetMapping("/invitations/{invitationId}/")
 	public String initAcceptInvitation(@RequestParam(value = "accepted", required = true) final String accepted,
-			final ModelMap model, @PathVariable("invitationId") final Integer invitationId) {
-		
-		// COMPROBAR QUE USUARIO NO HA ENTRADO POR URL (AUTENTICIDAD DE USUARIO, EVITAR QUE OTRO USUARIO ACEPTE LA INVITACION)
+			final ModelMap model, @PathVariable("invitationId") final Integer invitationId, Principal principal) {
+		Boolean hasPermission = false;
 		Invitation invitation = this.invitationService.findById(invitationId);
 		
-		if (invitation != null && accepted.equals("true")) {
+		if (invitation != null) {
+			hasPermission = principal.getName().equals(invitation.getReceiver().getUser().getUsername());
+			
+			if (!hasPermission) {
+				model.put("noPermission", true);
+			} 
+		} else {
+			model.put("noInvitation", true);
+		}
+			
+		if (hasPermission && invitation != null && accepted.equals("true")) {
+			
+//			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//			if (!authentication.getAuthorities().contains(new SimpleGrantedAuthority("premium"))) {
+//				return "redirect:/pay/championship/add?championshipId=" + championship.getId();
+//			}
 			
 			if (invitation.getTeam().getParticipants().size() < invitation.getTeam().getTeamSize()) {
 				Boolean isNotInChampionshipTeam =isNotInChampionshipTeam(invitation.getTeam().getChampionship().getId(), invitation.getReceiver());
@@ -167,12 +208,12 @@ public class InvitationController {
 				this.invitationService.delete(invitationId);
 			}
 			
-		} else if (invitation != null && accepted.equals("false")) {
+		} else if (hasPermission && invitation != null && accepted.equals("false")) {
 			this.invitationService.delete(invitationId);
 			model.put("notJoined", true);
 		} 
 		
-		Collection<Invitation> invitations = this.invitationService.findChampionshipInvitationsByUserName("Antonio");	// CAMBIAR AL NAME DEL PRINCIPAL CUANDO SE PUEDA
+		Collection<Invitation> invitations = this.invitationService.findChampionshipInvitationsByUsername(principal.getName());	
 		model.addAttribute("invitations", invitations);
 		model.addAttribute("areTeamInvitations", true);
 		return "invitations/listInvitations";
